+++ packages/components/psammead-brand/src/index.jsx
@@ -1,10 +1,10 @@
 import React, { Fragment } from 'react';
 import styled from 'styled-components';
-import { string, number, node, shape } from 'prop-types';
+import { string, number, node, shape, bool } from 'prop-types';
 import { C_POSTBOX, C_WHITE } from '@bbc/psammead-styles/colours';
 import VisuallyHiddenText from '@bbc/psammead-visually-hidden-text';
 import {
-  GEL_GROUP_3_SCREEN_WIDTH_MIN,
+  GEL_GROUP_2_SCREEN_WIDTH_MIN,
   GEL_GROUP_5_SCREEN_WIDTH_MIN,
 } from '@bbc/gel-foundations/breakpoints';
 import {
@@ -14,13 +14,15 @@ import {
   GEL_SPACING_TRPL,
 } from '@bbc/gel-foundations/spacings';
 
-const SVG_TOP_OFFSET_ABOVE_600PX = '1.75rem'; // 28px
-const SVG_BOTTOM_OFFSET_BELOW_600PX = '0.75rem'; // 12px
-const PADDING_AROUND_SVG_ABOVE_600PX = 56;
-const PADDING_AROUND_SVG_BELOW_600PX = 32;
+const SVG_TOP_OFFSET_ABOVE_400PX = '1.75rem'; // 28px
+const SVG_BOTTOM_OFFSET_BELOW_400PX = '0.75rem'; // 12px
+const PADDING_AROUND_SVG_ABOVE_400PX = 56;
+const PADDING_AROUND_SVG_BELOW_400PX = 32;
 
 const conditionallyRenderHeight = (svgHeight, padding) =>
   svgHeight ? `height: ${(svgHeight + padding) / 16}rem` : '';
+
+const TRANSPARENT_BORDER = `0.0625rem solid transparent`;
 
 const SvgWrapper = styled.div`
   max-width: ${GEL_GROUP_5_SCREEN_WIDTH_MIN};
@@ -30,15 +32,17 @@ const SvgWrapper = styled.div`
 const Banner = styled.div`
   background-color: ${C_POSTBOX};
   ${({ svgHeight }) =>
-    conditionallyRenderHeight(svgHeight, PADDING_AROUND_SVG_BELOW_600PX)};
+    conditionallyRenderHeight(svgHeight, PADDING_AROUND_SVG_BELOW_400PX)};
   width: 100%;
   padding: 0 ${GEL_SPACING};
 
-  @media (min-width: ${GEL_GROUP_3_SCREEN_WIDTH_MIN}) {
+  @media (min-width: ${GEL_GROUP_2_SCREEN_WIDTH_MIN}) {
     ${({ svgHeight }) =>
-      conditionallyRenderHeight(svgHeight, PADDING_AROUND_SVG_ABOVE_600PX)};
+      conditionallyRenderHeight(svgHeight, PADDING_AROUND_SVG_ABOVE_400PX)};
     padding: 0 ${GEL_SPACING_DBL};
   }
+  border-top: ${({ borderTop }) => borderTop && TRANSPARENT_BORDER};
+  border-bottom: ${({ borderBottom }) => borderBottom && TRANSPARENT_BORDER};
 `;
 
 const brandWidth = (minWidth, maxWidth) => `
@@ -56,12 +60,12 @@ const BrandSvg = styled.svg`
   box-sizing: content-box;
   fill: ${C_WHITE};
   padding-top: ${GEL_SPACING_DBL};
-  padding-bottom: ${SVG_BOTTOM_OFFSET_BELOW_600PX};
+  padding-bottom: ${SVG_BOTTOM_OFFSET_BELOW_400PX};
 
   ${({ maxWidth, minWidth }) => brandWidth(minWidth, maxWidth)}
 
-  @media (min-width: ${GEL_GROUP_3_SCREEN_WIDTH_MIN}) {
-    padding-top: ${SVG_TOP_OFFSET_ABOVE_600PX};
+  @media (min-width: ${GEL_GROUP_2_SCREEN_WIDTH_MIN}) {
+    padding-top: ${SVG_TOP_OFFSET_ABOVE_400PX};
     padding-bottom: ${GEL_SPACING_TRPL};
   }
 
@@ -153,35 +157,24 @@ StyledBrand.defaultProps = {
   serviceLocalisedName: null,
 };
 
-const Brand = ({
-  product,
-  serviceLocalisedName,
-  svgHeight,
-  minWidth,
-  maxWidth,
-  svg,
-  url,
-}) => {
-  const styledBrandProps = {
-    product,
-    serviceLocalisedName,
-    svgHeight,
-    minWidth,
-    maxWidth,
-    svg,
-  };
+const Brand = props => {
+  const { svgHeight, maxWidth, minWidth, url, borderTop, borderBottom } = props;
 
   return (
-    <Banner svgHeight={svgHeight}>
+    <Banner
+      svgHeight={svgHeight}
+      borderTop={borderTop}
+      borderBottom={borderBottom}
+    >
       {url ? (
         <SvgWrapper>
           <StyledLink href={url} maxWidth={maxWidth} minWidth={minWidth}>
-            <StyledBrand {...styledBrandProps} />
+            <StyledBrand {...props} />
           </StyledLink>
         </SvgWrapper>
       ) : (
         <SvgWrapper>
-          <StyledBrand {...styledBrandProps} />
+          <StyledBrand {...props} />
         </SvgWrapper>
       )}
     </Banner>
@@ -191,12 +184,16 @@ const Brand = ({
 Brand.defaultProps = {
   url: null,
   serviceLocalisedName: null,
+  borderTop: false,
+  borderBottom: false,
 };
 
 Brand.propTypes = {
   ...brandProps,
   url: string,
   serviceLocalisedName: string,
+  borderTop: bool,
+  borderBottom: bool,
 };
 
 export default Brand;